minor fix changes

+++ src/main/java/finance/life/monitoring/backend/feature/subscription/repository/SubscriptionRepository.java
@@ -4,6 +4,7 @@ import finance.life.monitoring.backend.feature.subscription.model.Subscription;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 import java.util.UUID;
 
@@ -12,5 +13,5 @@ public interface SubscriptionRepository extends JpaRepository<Subscription, UUID
     Optional<Subscription> getSubscriptionById(UUID uuid);
 
     @Query("SELECT SUM(s.amount) FROM Subscription s")
-    Optional<Float> getTotalOfSubscriptions();
+    Optional<BigDecimal> getTotalOfSubscriptions();
 }
